[transceivers] Fixed warning with ConsoleTransceiver

+++ src/shared/radio/ConsoleTransceiver/ConsoleTransceiver.h
@@ -36,7 +36,7 @@ public:
     bool send(uint8_t* pkt, size_t len)
     {
         auto serial = miosix::DefaultConsole::instance().get();
-        return serial->writeBlock(pkt, len, 0) == len;
+        return serial->writeBlock(pkt, len, 0) == static_cast<ssize_t>(len);
     }
 
     ssize_t receive(uint8_t* pkt, size_t len)
